Implemented Cluster initialization mechanism

+++ DistributedKvStore.Node/Controllers/DataController.cs
@@ -1,3 +1,4 @@
+using DistributedKvStore.Node.Services.Implementation.State;
 using DistributedKvStore.Node.Services.Interfaces;
 using DistributedKvStore.Shared.DTOs;
 using Microsoft.AspNetCore.Mvc;
@@ -9,15 +10,20 @@ namespace DistributedKvStore.Node.Controllers;
 public class DataController : ControllerBase
 {
     private readonly IKeyValueService _kvService;
+    private readonly INodeStateService _nodeState;
 
-    public DataController(IKeyValueService kvService)
+    public DataController(IKeyValueService kvService, INodeStateService nodeState)
     {
         _kvService = kvService;
+        _nodeState = nodeState;
     }
 
     [HttpGet("{key}")]
     public async Task<IActionResult> Get(string key)
     {
+        if (!_nodeState.IsInitialized)
+            return StatusCode(StatusCodes.Status503ServiceUnavailable, "Cluster/Node has not been Initialized yet.");
+
         var result = await _kvService.GetAsync(key);
         if (result == null)
             return NotFound();
@@ -28,6 +34,9 @@ public class DataController : ControllerBase
     [HttpPost]
     public async Task<IActionResult> Post([FromBody] KeyValueRequest request)
     {
+        if (!_nodeState.IsInitialized)
+            return StatusCode(StatusCodes.Status503ServiceUnavailable, "Cluster/Node has not been Initialized yet.");
+
         await _kvService.PutAsync(request.Key, request.Value);
         return Ok();
     }
@@ -35,6 +44,9 @@ public class DataController : ControllerBase
     [HttpPut]
     public async Task<IActionResult> Put([FromBody] KeyValueRequest request)
     {
+        if (!_nodeState.IsInitialized)
+            return StatusCode(StatusCodes.Status503ServiceUnavailable, "Cluster/Node has not been Initialized yet.");
+
         await _kvService.UpdateAsync(request.Key, request.Value);
         return Ok();
     }
@@ -42,6 +54,9 @@ public class DataController : ControllerBase
     [HttpDelete("{key}")]
     public async Task<IActionResult> Delete(string key)
     {
+        if (!_nodeState.IsInitialized)
+            return StatusCode(StatusCodes.Status503ServiceUnavailable, "Cluster/Node has not been Initialized yet.");
+
         await _kvService.DeleteAsync(key);
         return Ok();
     }
